move the helios code to its own file
